Refactored Process Checker

+++ Tribler/Core/Modules/process_checker.py
@@ -1,4 +1,5 @@
 import os
+import psutil
 
 from Tribler.Core.Config.tribler_config import TriblerConfig
 
@@ -10,37 +11,29 @@ class ProcessChecker(object):
     """
     This class contains code to check whether a Tribler process is already running.
     """
-
     def __init__(self, state_directory=None):
-        """
-        Check whether a lock file exists in the Tribler directory. If not, create the file. If it exists,
-        check the PID that is written inside the lock file.
-        """
-        self.already_running = False
-
         self.state_directory = state_directory or TriblerConfig().get_state_dir()
         self.lock_file_path = os.path.join(self.state_directory, LOCK_FILE_NAME)
 
         if os.path.exists(self.lock_file_path):
-            try:
-                file_pid = int(self.get_pid_from_lock_file())
-            except ValueError:
-                # Apparently, the data written inside the lock file is not an int, just remove the file and recreate it.
+            # Check for stale lock file (created before the os was last restarted).
+            # The stale file might contain the pid of another running process and
+            # not the Tribler itself. To find out we can simply check if the lock file
+            # was last modified before os reboot.
+            # lock_file_modification_time < system boot time
+            file_pid = self.get_pid_from_lock_file()
+            if file_pid < 1 or os.path.getmtime(self.lock_file_path) < psutil.boot_time():
                 self.remove_lock_file()
-                self.create_lock_file()
-                return
 
-            if file_pid == os.getpid():
-                # Ignore when we find our own PID inside the lock file
-                self.already_running = False
-            elif file_pid != os.getpid() and not ProcessChecker.is_pid_running(file_pid):
-                # The process ID written inside the lock file is old, just remove the lock file and create a new one.
-                self.remove_lock_file()
-                self.create_lock_file()
-            else:
-                self.already_running = True
-        else:
-            self.create_lock_file()
+        self.already_running = self.is_process_running()
+
+    def is_process_running(self):
+        if os.path.exists(self.lock_file_path):
+            file_pid = self.get_pid_from_lock_file()
+
+            if file_pid == os.getpid() or ProcessChecker.is_pid_running(file_pid):
+                return True
+        return False
 
     @staticmethod
     def is_pid_running(pid):
@@ -64,18 +57,27 @@ class ProcessChecker(object):
         if not os.path.exists(self.state_directory):
             os.makedirs(self.state_directory)
 
+        # Remove the previous lock file
+        self.remove_lock_file()
+
         with open(self.lock_file_path, 'wb') as lock_file:
             lock_file.write(str(os.getpid()))
 
     def remove_lock_file(self):
         """
-        Remove the lock file.
+        Remove the lock file if it exists.
         """
-        os.unlink(self.lock_file_path)
+        if os.path.exists(self.lock_file_path):
+            os.unlink(self.lock_file_path)
 
     def get_pid_from_lock_file(self):
         """
         Returns the PID from the lock file.
         """
+        if not os.path.exists(self.lock_file_path):
+            return -1
         with open(self.lock_file_path, 'rb') as lock_file:
-            return lock_file.read()
+            try:
+                return int(lock_file.read())
+            except ValueError:
+                return -1
